Skip dbname's we can't access
The datadog user may only have access to the databases which the customer wishes
to monitor. Without this patch, a failure to invoke dbstats on _any_ database
would raise an exception and prevent monitoring for _all_ databases. With this
patch, databases that cannot be accessed are ignored with a warning.