[http_check] Fix content matching for non-ascii characters
As stated in pyyaml's docs (http://pyyaml.org/wiki/PyYAMLDocumentation#StringconversionPython2only)
yaml strings are converted to `unicode` python objects when they
contain non-ascii characters. The existing test would use the
type `str` for `content_match` instead so it passed but the
check actually didn't work  with non-ascii chars in `content_match`

The reason it didn't work is that it'd compare a `unicode` object
with the undecoded `content` of the response.

To fix this, change the test to use a `unicode` object instead, and
make the check use the response's `text` (i.e. the decoded `content`)
when `content_match` is a `unicode` object.