[win32][gui] Fix old process detection and termination
The agent manager would crash on startup if the stored pid was
referring to a process that the user couldn't access or terminate.

To fix this, ignore "access denied" errors.

Also, switch to psutil for simplicity. I think we didn't use psutil
before because it still had some bugs on Windows, but it looks like
these bugs have been fixed since then.