[docker_daemon] Report as many cgroup metrics as possible
If one cgroup stat file is missing, still try to report the metrics
from the other files.

Also, remove the retry logic as it seems to have been made at a time
when `find_cgroup_filename_pattern` was used. Since we're trying
all the stat files now we are in fact retrying `len(CGROUP_METRICS)`
times