[utils/hostname] better EC2 detection
If the hostname is None, we should try to get the EC2 instance-id.
This was reported by Atlassian, and they proposed this fix with a
detailed error report.
The hostname can be None when `/bin/hostname -f` returns an error, which
can happen at AMI creation when the hosts file is not ready.