[forwarder] Recompute next flush of transactions on flush timeout
This is needed so that when a flush timeout occurs, the transactions
of that flush do not stay at the head of the queue (if they do stay
at the head of the queue then, when the max queue size is hit, newer
transactions are deleted first, which is not desirable).

Recomputing the next flush time might delay these transactions
slightly (up to MAX_WAIT_FOR_REPLAY seconds, i.e. 90 seconds), but it's
ok here since at this point these transactions have probably been
waiting for some time already.