[check] Fix misattribution of check instance warning
If a check has multiple instances, and the first instance adds a
warning and then raises an exception, the next instance of the check
would actually have that warning in its status (even if this second
instance ran fine).

Fix this by always discarding remaining warnings at the end of each
instance run.

Also, add a test reproducing the issue.