[dogstatsd] Parse distribution metrics and ignore them
Instead of failing on known but unsupported metric types, ignore these
metrics and continue parsing the rest of the packet.

Without this, sending a new "distribution" type metric in a datagram
makes dogstatsd drop the whole datagram.

This change has virtually no impact on the performance of the dogstatsd
parser.