[init scripts] Fix agent invocations when dd-agent has no login shell
Agent 6 creates the dd-agent user with no login shell. So, if a user
install Agent 6 from scratch, and then downgrades to Agent 5, the
dd-agent still has no login shell set. Therefore we have to make all
agent invocations work with no login shell, this fixes.

We don't make the Agent 5 create the dd-agent user with no login shell
though, this is a change that may break some setups, so we want to make
it with the major v6 version.